+++ backend/main.py
@@ -7,6 +7,7 @@ from fastapi.staticfiles import StaticFiles
 
 from routers.apikeys import router as apikeys_router
 from routers.map import router as map_router
+from routers.mapdefaults import router as mapdefaults_router
 from routers.nominatim import router as nominatim_router
 from routers.settings import router as settings_router
 from routers.taxonomy import router as taxonomy_router
@@ -26,6 +27,7 @@ app.add_middleware(
 
 app.include_router(apikeys_router)
 app.include_router(map_router)
+app.include_router(mapdefaults_router)
 app.include_router(weather_router)
 app.include_router(version_router)
 app.include_router(nominatim_router)